refactor(typescript): rename types
