+++ client/src/components/layout/BlogItem.tsx
@@ -1,5 +1,6 @@
 import stitches from '../../stitches.config';
 import { Link } from 'react-router-dom';
+import IBlog from '../../types/IBlog';
 
 const { styled } = stitches;
