--- client/src/types/authentication.ts
@@ -1,17 +1,0 @@
-interface signUpResponse {
-  success: true;
-  message?: string;
-}
-
-interface loginResponse {
-  success: boolean;
-  message?: string;
-  token: string;
-}
-
-interface IUserAuth {
-  isloggedIn: boolean;
-  token: string;
-}
-
-export type { signUpResponse, loginResponse, IUserAuth };
